Add binary sizes to stress test metrics
